+++ tests/data/functional_tests/small/load.sql
@@ -1,0 +1,1 @@
+COPY drivers FROM 'small/drivers.csv' (FORMAT 'csv', quote '"', delimiter ',', header 0);