+++ tests/data/functional_tests/tiny/load.sql
@@ -1,0 +1,1 @@
+COPY foo FROM 'tiny/foo.csv' (FORMAT 'csv', quote '"', delimiter ',', header 0);